printf() * variable (as in "%.*s") must be int
git-svn-id: https://svn.apache.org/repos/asf/apr/apr/trunk@62059 13f79535-47bb-0310-9956-ffa450edef68

+++ test/aprtest.h
@@ -62,7 +62,7 @@
 
 #define APR_TEST_BEGIN(rv, desc, op) \
     fprintf(stdout, "%s%.*s ", APR_TEST_PREFIX desc,                  \
-            strlen(desc) < 37 ? 40 - strlen(desc) : 3,                \
+            strlen(desc) < 37 ? (int)(40 - strlen(desc)) : 3,         \
             "........................................");              \
     APR_TEST_MORE(rv, op)
